Added DBG macro in util.h to facilitate debugging
This macro allows debugging statements (typically printf's or cout's) to
be activated or deactivated with a single comment.  Uncomment the line:
//#define ENABLE_DASH_DEBUG
in util.h to enable debugging statements.

When commented any code wrapped with the DBG() macro will simply be removed
by the preprocessor.  When not commented all such wrapped statements will
be present.

For maximum effectiveness it is best that util.h be the first effective include
in all source files.  It is also possible to enable the macro for a single file
by temporarily adding #define ENABLE_DASH_DEBUG to the top of the file.

Code committed to non-development branches should always have the define
commented.